extract types.ts from panels/filters.ts
Pull DateField, TagFilter, DateFilter, Filter, and FilterState into
panels/filters/types.ts. Re-export them from panels/filters.ts so
existing call sites are unchanged. First step of the four-way split
documented in panels/AGENTS.md.
